Improve QR code validation by specifying data type for homography matrix
Update `validate_slot_qrs.py` to explicitly cast the homography matrix to `np.float32` to resolve an OpenCV assertion error when applying perspective transformation.

+++ python/validate_slot_qrs.py
@@ -111,7 +111,7 @@ def validate_slot_qrs(camera_index, resolution, homography_matrix, expected_slot
             print(f"Skipping undistortion (all coefficients are zero)", file=sys.stderr)
     
     # Apply homography transformation to get rectified view
-    H = np.array(homography_matrix).reshape(3, 3)
+    H = np.array(homography_matrix, dtype=np.float32).reshape(3, 3)
     h, w = frame.shape[:2]
     rectified = cv2.warpPerspective(frame, H, (w, h))
